+++ config/initializers/action_mailer.rb
@@ -1,5 +1,9 @@
-ActionMailer::Base.default_url_options[:host] = ENV['website_url'].to_s.gsub(/^https?:\/\//, '')
-unless Rails.env.test?
+ActionMailer::Base.default_url_options[:host] = ENV['website_url'].to_s.gsub(%r{^https?://}, '')
+if Rails.env.test?
+  ActionMailer::Base.delivery_method = :test
+  ActionMailer::Base.smtp_settings[:email] = 'travis-ci@example.com'
+  EMAILS_ENABLED = true
+else
   ActionMailer::Base.delivery_method = :smtp
   ActionMailer::Base.smtp_settings = {
     enable_starttls_auto: true,
@@ -11,6 +15,5 @@ unless Rails.env.test?
     user_name: ENV['smtp_user_name'],
     password: ENV['smtp_password']
   }
-else
-  ActionMailer::Base.smtp_settings[:email] = "travis-ci@example.com"
+  EMAILS_ENABLED = (ENV['emails_enabled'] == 'true')
 end